+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   connect to peers that do not support wire-protocol version 1, which is supported
   since version 1.1.0.
 - The macOS installer no longer overwrites the service files when reinstalling.
+- Cache smart contract modules on startup from existing state to improve smart
+  contract execution.
 
 ## concordium-node 3.0.0
 - Fix a bug due to incorrect use of LMDB database environments where a node